Search for key by CKA_ID of the certificate we found
http://david.woodhou.se/draft-woodhouse-cert-best-practice.html#rfc.section.8.2
tells us that when we're searching for a key based on a URI that
successfully identified a certificate, then if the key is not found and
the original search was by CKA_LABEL of the certificate, we should
repeat the search using the CKA_ID of the certificate that was actually
found, but not requiring a CKA_LABEL match.

This covers the case when a PIV user specifies something like
 pkcs11:manufacturer=piv_II;object=Certificate%20for%20Digital%20Signature
(which is likely to be the kind of thing that GUI selection tools will
provide when allowing the user to select a cert).