Remove superfluous 'single' argument from getCertsInSession()
It doesn't actually do anything.

I think I originally intended it to return {cert,id,label} only for
the first cert, or a []cert array for a list. But that's icky, so
I ended up just defining the CertObjInfo and returning that. Partly
because the read-certificate-data *should* be printing the URI of
the certs it lists, so we do want to propagate id/label for those
too.