Use make to build all certs
Having the Makefile import the certs into the SoftHSM token before running
'go test' doesn't work very well if said certs are only generated later
when 'generate-certs.sh' is invoked from the test run itself.

So move it all to the Makefile.

We do use a GNU make'ism here just to generate the permutations of
key sizes/curves and digests, but we could expand that manually if
we wanted to.