Support password-protected TPM signing keys
* Only password-protected TPM signing keys are supported. It isn't
currently possible to use passwords for parents or hierarchies.
* Modify the README appropriately, to convey the above.
* Refactor testing and Makefile, so that TPM tests are in their own
file, and so that they can be run in isolation (without having to set
up SoftHSM, for example).
* Change RSA signing, so that only keys with the sign capability can be
used to perform signatures (note that this is yet to be tested since
it may not be possible to create such keys using the create_tpm2_key
utility.